Fixed plo_dot_prod_q32_parallel
The data is split amongst the workers, such that every core (except
the last one) will process an even number of elements. The reason for
this design choice is so that the result of `plp_dot_prod_q32` is
identical to `plp_dot_prod_q32_parallel`. We need to always compute
the MAC of two elements, sum the up and then do the round / norm
operation.