Test Framework: added InplaceArgument
The problem with InplaceArguments is that they need to be reset every
time the benchmark is executed. This is because the funciton might be
data-dependent. Here, we generate exactly three arrays for each
InplaceArgument, one is for the original data, one is for the expected
result, and one is for the working data. Only the working data can be
put on L1, and both others are always on L2 to store memory.

The README was updated accordingly.