Test Framework: Removed memory requirements
The test framework now works no matter the number of sweep iterations.
It approximates the memory requirements for all sweep iterations (test
cases) and generates multiple test projects, where the data fits into L2
memory.

As a side effect, `plptest` timeout does no longer work (bug is
reported). As alternative, we use the linux program `timeout`. Now, the
framework parses the output of the compilation and of the execution, and
can report what the problem was when a timeout would occurr. The
previous behavior was that it just reports `KO`, and the case of an
error in the compilation was indistinguishable to an invalid access (or
an endless loop). This now has hanged.